Update file structure and update paths
