--- src/components/ArrayButton.jsx
@@ -1,7 +1,0 @@
-const ArrayButton = (props) => {
-    return (
-        <button>{props.name} {props.length}</button>
-    )
-}
-
-export default ArrayButton;
